Created "http" library, moved platform-dependent utilities to "utility" library
