+++ lib/basic/CMakeLists.txt
@@ -15,8 +15,6 @@ add_library(${PROJECT_NAME}
     "src/basic_parser.c"
     "src/basic_program.c"
     "src/basic_runner.c"
-    "src/utils.c"
-    "src/utils.h"
 )
 
 add_library(${PROJECT_NAME}::${PROJECT_NAME} ALIAS ${PROJECT_NAME})
@@ -24,12 +22,13 @@ add_library(${PROJECT_NAME}::${PROJECT_NAME} ALIAS ${PROJECT_NAME})
 # Add some preprocessor defines
 target_compile_definitions(${PROJECT_NAME} PUBLIC "$<$<CONFIG:DEBUG>:DEBUG>")
 
-add_dependencies(${PROJECT_NAME} basic_system_interface data_structures)
+add_dependencies(${PROJECT_NAME} basic_system_interface data_structures utility)
 
 target_link_libraries(${PROJECT_NAME}
     PRIVATE
         basic_system_interface::basic_system_interface
         data_structures::data_structures
+        utility::utility
 )
 
 target_include_directories(${PROJECT_NAME} PUBLIC
