+++ lib/basic/src/basic_lexer.c
@@ -1,6 +1,6 @@
-
 #include "basic/basic.h"
-#include "utils.h"
+
+#include <utility/utils.h>
 
 #include <stdio.h>
 #include <stdlib.h>
