+++ lib/basic/src/basic_parser.c
@@ -1,9 +1,8 @@
-
-#include "utils.h"
 #include "basic/basic.h"
 
 #include <data_structures/stack.h>
 #include <data_structures/queue.h>
+#include <utility/utils.h>
 
 #include <stdio.h>
 #include <stdlib.h>
