+++ lib/basic/src/basic_runner.c
@@ -1,9 +1,9 @@
 #include <data_structures/stack.h>
 #include <basic_system_interface/system.h>
+#include <utility/utils.h>
 
 #include "basic/basic.h"
 #include "basic/ast.h"
-#include "utils.h"
 
 #define _USE_MATH_DEFINES
