+++ lib/utility/src/platform.c
@@ -1,4 +1,4 @@
-#include "platform.h"
+#include "utility/platform/platform.h"
 
 #ifdef __linux__
 #include <unistd.h>
