--- src/utils.c
@@ -1,1 +1,0 @@
-#include "utils.h"
